requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.1
paho-mqtt==2.1.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
uritemplate==4.2.0
zope.interface==8.0
requests==2.32.5